core-service/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
supabase==2.28.3
pydantic-settings==2.6.1
PyJWT==2.12.0
bcrypt==4.2.1
slowapi==0.1.9
httpx==0.27.2
